Separates main.rs from lib.rs.
